feat/order-create: 주문 등록 테스트 코드

+++ src/main/java/com/stockexchange/domain/order/dto/OrderDetailResDTO.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Data
@@ -15,7 +16,7 @@ public class OrderDetailResDTO {
     private Long stockId;
     private Long orderId;
     private int orderCount;
-    private long orderPrice;
+    private BigDecimal orderPrice;
     private OrderType orderType;
     private OrderStatus orderStatus;
     private int orderRemainCount;
